backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]
pydantic>=2.0
pydantic-settings
python-multipart
supabase
sqlalchemy>=2.0
alembic
psycopg2-binary
google-generativeai
python-dotenv
openai
requests
httpx
resend
geopy
